feat!: Upgrade Spring Framework
OCD-4237

+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/dao/ListingGraphDAO.java
@@ -1,10 +1,7 @@
 package gov.healthit.chpl.dao;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
-
-import javax.persistence.Query;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
@@ -16,6 +13,7 @@ import gov.healthit.chpl.entity.listing.CertifiedProductEntity;
 import gov.healthit.chpl.entity.listing.ListingToListingMapEntity;
 import gov.healthit.chpl.exception.EntityCreationException;
 import gov.healthit.chpl.util.ErrorMessageUtil;
+import jakarta.persistence.Query;
 import lombok.extern.log4j.Log4j2;
 
 @Repository(value = "listingGraphDao")
@@ -80,9 +78,9 @@ public class ListingGraphDAO extends BaseDAOImpl {
                                 + "FROM CertifiedProductEntity listing "
                                 + "WHERE listing.id IN (:listingIds) "
                                 + "AND listing.deleted <> true",
-                        BigDecimal.class);
+                        Double.class);
         query.setParameter("listingIds", listingIds);
-        BigDecimal result = (BigDecimal) query.getSingleResult();
+        Double result = (Double) query.getSingleResult();
         if (result == null) {
             return null;
         }
